requirements.txt
google-genai
requests
# --- RAG ve Hafıza ---
chromadb>=0.4.0
sentence-transformers>=2.2.0
torch>=2.0.0
# --- Yardımcılar ---
tqdm  # İndeksleme sırasında progress bar için (opsiyonel ama iyi pratik)